Handling User Not Found

+++ internal/app/repository/user.repository.go
@@ -45,14 +45,22 @@ func (u *UserRepository) GetUserDataByDocRefID(ctx context.Context, docRefID str
 }
 
 // TODO handle collection not found
+// TODO NOTE : Using Snapshot To Validate
 func (u *UserRepository) DeleteUserDataByDocRefID(ctx context.Context, docRefID string) (bool, error) {
 	docRef := u.DB.FirestoreClient.Collection(USER_COLLECTION_NAME).Doc(docRefID)
-	_, err := docRef.Delete(ctx)
+	snapshot, err := docRef.Get(ctx)
 	if err != nil {
-		return false, err
+		return false, nil
 	}
+	if snapshot.Exists() {
+		_, err := docRef.Delete(ctx)
+		if err != nil {
+			return false, err
+		}
+		return true, nil
+	}
+	return false, nil
 
-	return true, nil
 }
 
 func (u *UserRepository) GetAllUserData(ctx context.Context) ([]entity.User, error) {
